Added significant testing coverage
