+++ package.json
@@ -7,6 +7,9 @@
   "engines": {
     "node": ">=6.13.0"
   },
+  "jest": {
+      "setupFiles": [ "./setupTests.js"]
+  },
   "scripts": {
     "server": "nodemon server/index.js",
     "build": "webpack -d --watch",
